sprite conform to norm
